# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ChiefVoice CRM API
  version: 1.0.0
  description: |
    AI-powered voice CRM backend.

    **Authentication:** All protected endpoints require a identity-provider-issued ID token.
    Pass it as `Authorization: Bearer <token>`.
    The SSE endpoint `/api/logs-stream` uses a short-lived ticket created by `/api/logs-stream/ticket`
    because browser EventSource cannot set custom Authorization headers. Other
    protected endpoints do not accept JWTs in query strings.

    **Roles:** `Organization Admin`, `Super Admin`, `Agent`, `Manager`.
    Platform-admin endpoints additionally require the email to be in `PLATFORM_ADMIN_EMAILS`.

  contact:
    name: ChiefVoice Support
    email: support@chiefvoice.ai
servers:
  - url: /api
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication ID token
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
...
    Product:
      type: object
      properties:
        id:    { type: string }
        name:  { type: string }
        brand: { type: string }
        unit:  { type: string }
        price: { type: number }
        stock: { type: integer }
    Order:
      type: object
      properties:
        id:       { type: string }
        customer: { type: string }
        total:    { type: number }
        status:   { type: string }
        source:   { type: string }
    Customer:
      type: object
      properties:
        id:       { type: string }
        name:     { type: string }
        phone:    { type: string }
        type:     { type: string }
        locality: { type: string }
        ltv:      { type: number }
        khata:    { type: number }
...
    Channel:
      type: object
      properties:
        id:       { type: string }
        platform: { type: string, enum: [whatsapp, instagram] }
        accountId: { type: string }
        status:   { type: string }

# ── Global security default ───────────────────────────────────────────────────
security:
  - BearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /lead-responses:
    get:
      tags: [Leads]
      summary: List lead call responses
      responses:
        '200': { description: Lead responses }
    post:
      tags: [Leads]
      summary: Create lead response
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Created }
...
  /catalog:
    get:
      tags: [Catalog]
      summary: List products
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Product' } }
    post:
      tags: [Catalog]
      summary: Add product
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        '201': { description: Created }
  /catalog/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Catalog]
      summary: Update product
      responses:
        '200': { description: Updated }
  /orders:
    get:
      tags: [Catalog]
      summary: List orders
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Order' } }
    post:
      tags: [Catalog]
      summary: Create order
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Order' }
      responses:
        '201': { description: Created }
  /orders/{id}/status:
    patch:
      tags: [Catalog]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
      responses:
        '200': { description: Updated }
  /customers:
    get:
      tags: [Catalog]
      summary: List customers
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Customer' } }
    post:
      tags: [Catalog]
      summary: Create customer
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Customer' }
      responses:
        '201': { description: Created }
  /customers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Catalog]
      summary: Get customer
      responses:
        '200': { description: Customer }
        '404': { $ref: '#/components/responses/NotFound' }
  /khata/{customerId}/payment:
    post:
      tags: [Catalog]
      summary: Record khata payment
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number }
                note:   { type: string }
      responses:
        '200': { description: Recorded }
  /khata/{customerId}/reminder:
    post:
      tags: [Catalog]
      summary: Send khata payment reminder
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Reminder sent }
  /questions:
    get:
      tags: [Catalog]
      summary: Get AI qualification questions
      responses:
        '200': { description: Questions }
    post:
      tags: [Catalog]
      summary: Replace qualification questions
      requestBody:
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200': { description: Saved }
  /v2/chroma/search:
    get:
      tags: [Catalog]
      summary: Semantic search in Chroma knowledge store
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Results }

  # ── Channels ──────────────────────────────────────────────────────────────
  /channels:
    get:
      tags: [Channels]
      summary: List connected channels
      responses:
        '200':
          description: Channels
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Channel' } }
    post:
      tags: [Channels]
      summary: Connect a new channel
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Channel' }
      responses:
        '201': { description: Created }
  /channels/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Channels]
      summary: Update channel
      responses:
        '200': { description: Updated }
    delete:
      tags: [Channels]
      summary: Delete channel
      responses:
        '200': { description: Deleted }
...